feat: `AlertToast` and `useToast` hook
