+++ packages/react-styled-ui/src/index.js
@@ -1,4 +1,5 @@
 import Alert from './Alert';
+import AlertToast from './AlertToast';
 import Badge from './Badge';
 import Box from './Box';
 import Button from './Button';
@@ -52,6 +53,7 @@ import withTheme from './withTheme';
 
 export {
   Alert,
+  AlertToast,
   Badge,
   Box,
   Button,
