+++ packages/styled-ui-docs/pages/alert.mdx
@@ -168,7 +168,7 @@ Setting the `icon` prop to `false` will remove the icon altogether.
 
 ## Props
 
-`Alert` composes the `Flex` component.
+`Alert` composes the [`Flex`](./flex) component.
 
 | Name | Type | Default | Description |
 | :--- | :--- | :------ | :---------- |
